ref(catalog): move testing.py into catalog/tests/
testing.py contains only test utilities (Textual Pilot helpers) and is
not part of the public catalog API — it belongs alongside the tests that
use it.
